requirements.txt
fastapi
uvicorn
faiss-cpu
pyarrow
lxml
sentence-transformers
pydantic